Got DataBlobArgs function working.

+++ Pulsar4X/Pulsar4X.ECSLib/DataBlobs/AbilityDBs/FuelStorageDB.cs
@@ -13,9 +13,9 @@ namespace Pulsar4X.ECSLib
         public int FuelStorage { get { return _fuelStorage; } internal set { _fuelStorage = value; } }
 
 
-        public FuelStorageDB(int fuelStorage)
+        public FuelStorageDB(double fuelStorage)
         {
-            _fuelStorage = fuelStorage;
+            _fuelStorage = (int)fuelStorage;
         }
 
         public FuelStorageDB(FuelStorageDB db)
